requirements.txt
pytest-8.2.2
selenium-4.22.0